requirements.txt
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
pywin32==224
